Reorganize function locations
